feat: implement forgot and reset password functionality
