--- .husky/pre-commit
@@ -1,1 +1,0 @@
-npm run lint-staged --concurrent false
