+++ src/auth/dto/auth-credentials.dto.ts
@@ -34,3 +34,15 @@ export class AuthSignInDto {
   @IsPassword()
   password: string;
 }
+
+export class ForgotPasswordDto {
+  @IsEmail()
+  email: string;
+}
+
+export class ResetPasswordDto {
+  token: string;
+
+  @IsPassword()
+  newPassword: string;
+}
