+++ src/auth/jwt.strategy.ts
@@ -2,6 +2,7 @@ import { Injectable, UnauthorizedException } from "@nestjs/common";
 import { ConfigService } from "@nestjs/config";
 import { PassportStrategy } from "@nestjs/passport";
 import { ExtractJwt, Strategy } from "passport-jwt";
+
 import { JwtPayload } from "./jwt-payload.interface";
 import { User } from "./schema";
 import UsersRepository from "./user.repository";
