+++ src/profile/dto/profile.dto.ts
@@ -1,8 +1,4 @@
-import {
-  IsEmail,
-  IsName,
-  IsPassword,
-} from "@/common/decorators/dto.decorators";
+import { IsEmail, IsName } from "@/common/decorators/dto.decorators";
 
 export class UpdateProfileDto {
   @IsName()
@@ -10,16 +6,4 @@ export class UpdateProfileDto {
 
   @IsEmail()
   email: string;
-}
-
-export class ForgotPasswordDto {
-  @IsEmail()
-  email: string;
-}
-
-export class ResetPasswordDto {
-  token: string;
-
-  @IsPassword()
-  newPassword: string;
 }
