+++ src/profile/profile.module.ts
@@ -1,8 +1,8 @@
 import { Module } from "@nestjs/common";
+import { ConfigModule } from "@nestjs/config";
 
 import AuthModule from "@/auth/auth.module";
 import { CommonModule } from "@/common/common.module";
-import EmailService from "@/common/services/email.service";
 import DatabaseModule from "@/database/database.module";
 
 import ProfileController from "./profile.controller";
@@ -10,9 +10,9 @@ import ProfilesRepository from "./profile.repository";
 import ProfilesService from "./profile.service";
 
 @Module({
-  imports: [AuthModule, DatabaseModule, CommonModule],
+  imports: [AuthModule, DatabaseModule, CommonModule, ConfigModule],
   controllers: [ProfileController],
-  providers: [ProfilesService, ProfilesRepository, EmailService],
+  providers: [ProfilesService, ProfilesRepository],
   exports: [ProfilesRepository],
 })
 export default class ReviewsModule {}
